Add fix to handle errors for Outlook
On a new computer you cannot klick on links without getting
errormessages. Microsofts response is to reinstall Internet Explorer.
but there is a fix for this and this script changes the relevant
registry keys for this to function.

In this case Firefox is used as default browser for HTML-files.

Not tested on clean machines but should do the job.